refatoracao

+++ EstatisticasMundial+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  EstatisticasMundial+CoreDataProperties.swift
 //  projetoCovidMVVM
 //
-//  Created by Roberto Edgar Geiss on 01/05/23.
+//  Created by Roberto Edgar Geiss on 02/05/23.
 //
 //
 
@@ -20,7 +20,7 @@ extension EstatisticasMundial {
     @NSManaged public var cases: Int32
     @NSManaged public var critical: Int32
     @NSManaged public var deaths: Int32
-    @NSManaged public var population: Int32
+    @NSManaged public var population: Int64
     @NSManaged public var recovered: Int32
     @NSManaged public var todayCases: Int32
     @NSManaged public var todayRecovered: Int32
